feat: add basic auth & props to oauth2
