+++ packages/pieces/src/lib/framework/property/base-prop.ts
@@ -12,9 +12,9 @@ export enum PropertyType {
 	CHECKBOX = 'CHECKBOX',
 	OAUTH2 = 'OAUTH2',
 	SECRET_TEXT = 'SECRET_TEXT',
-	CUSTOM_AUTH = 'CUSTOM_AUTH',
 	ARRAY = 'ARRAY',
 	OBJECT = 'OBJECT',
+	BASIC_AUTH = "BASIC_AUTH",
 	JSON = 'JSON',
 }
